:recycle: 알람창 수정
- 취소가 가능한 알람창으로 수정

+++ frontend/src/components/NoteList.vue
@@ -109,26 +109,29 @@ export default {
 
     clickDelete(event, noteId) {
       event.stopPropagation();
+      const result = window.confirm('정말 삭제하시겠습니까 ?');
 
-      axios.defaults.withCredentials = true;
-      axios.delete(`http://localhost:8080/api/notes/${noteId}`)
-      .then(() => {
-        alert('정말 삭제하시겠습니까?');
-        window.location.href = '/notes';
-      }).catch(error => {
-        const errorStatus = error.response.status;
+      if (result) {
+        axios.defaults.withCredentials = true;
+        axios.delete(`http://localhost:8080/api/notes/${noteId}`)
+        .then(() => {
+          alert('정말 삭제하시겠습니까?');
+          window.location.href = '/notes';
+        }).catch(error => {
+          const errorStatus = error.response.status;
 
-        if (errorStatus === 401) {
-          localStorage.removeItem('vuex');
-          alert('로그인이 필요합니다.');
-          window.location.href = '/';
+          if (errorStatus === 401) {
+            localStorage.removeItem('vuex');
+            alert('로그인이 필요합니다.');
+            window.location.href = '/';
 
-        } else if (errorStatus === 400) {
-          const errorMessage = error.response.data.message;
-          alert(errorMessage);
-          this.$router.push('/');
-        }
-      })
+          } else if (errorStatus === 400) {
+            const errorMessage = error.response.data.message;
+            alert(errorMessage);
+            this.$router.push('/');
+          }
+        })
+      }
     },
   },
 }
